Try to make the API methods available slightly more sane
The core encryption/decryption functions (Encrypt/Decrypt) are now in-place-only.  If you want to decrypt into a different buffer than your source, then copy the data to the destination buffer and encrypt it there.  The original code destroyed the source buffer as part of processing _anyway_, so those overloads were dangerous to the uninitiated anyway.

Furthermore, the string operations now take ReadOnlySpan<char>, so they'll work on strings as well as character arrays.